feat: hook up help to login and logout commands
