+++ src/commands/logout.ts
@@ -5,18 +5,17 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 
 import SfCommand from '../sf-command';
 import { logout } from '../utils';
 
+Messages.importMessagesDirectory(__dirname);
+const messages = Messages.loadMessages('@salesforce/sf-demo', 'logout');
+
 export default class Logout extends SfCommand {
-  public static description = `logout of a Salesforce account or enviornment
-
-  Logout of all accounts or a specific user.
-  `;
-
-  public static examples = ['sf logout', 'sf logout [user]'];
+  public static summary = messages.getMessage('summary');
 
   public async run(): Promise<AnyJson> {
     const { args } = await this.parse(Logout);
